# FrontEnd/SakuraEmailAngular/swagger/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Sakura.EmailContact",
    "version": "v1"
  },
  "paths": {
    "/api/Campaigns": {
      "post": {
        "tags": ["Campaigns"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCampaignDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCampaignDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddCampaignDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Contacts": {
      "post": {
        "tags": ["Contacts"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddContactDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddContactDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddContactDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ContactDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactDto"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Contacts"],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContactDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContactDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContactDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Contacts/lists": {
      "post": {
        "tags": ["Contacts"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddContactListDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddContactListDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddContactListDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ContactListDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactListDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactListDto"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Contacts"],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContactListDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContactListDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContactListDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Contacts/lists/{listId}/contacts": {
      "post": {
        "tags": ["Contacts"],
        "parameters": [
          {
            "name": "listId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int32"
                },
                "nullable": true
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int32"
                },
                "nullable": true
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int32"
                },
                "nullable": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ContactListDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactListDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactListDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Contacts/lists/{id}": {
      "get": {
        "tags": ["Contacts"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ContactListDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactListDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactListDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AddEventDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "AddCampaignDto": {
        "required": ["emailTemplateId", "name"],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 150,
            "minLength": 4,
            "type": "string"
          },
          "emailTemplateId": {
            "type": "integer",
            "format": "int32"
          },
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AddEventDto"
            },
            "nullable": true
          },
          "contactLists": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "EventDto": {
        "type": "object",
        "properties": {
          "scheduleJobId": {
            "type": "string",
            "nullable": true
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "hour": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "ContactDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ContactListDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "contacts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContactDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CampaignDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventDto"
            },
            "nullable": true
          },
          "contactLists": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContactListDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AddContactDto": {
        "required": ["email", "name"],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          }
        },
        "additionalProperties": false
      },
      "AddContactListDto": {
        "required": ["name"],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "contactIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}